projects/capstone/starter/requirements.txt
alembic==1.6.5
appdirs==1.4.4
argon2-cffi==21.1.0
astroid==2.2.5
attrs==21.2.0
Babel==2.9.0
backcall==0.2.0
bleach==4.1.0
blinker==1.4
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
Click==7.0
colorama==0.4.4
cryptography==3.3.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.0.0
Flask-Moment==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gunicorn==20.0.4
Jinja2==2.10.1
jsonschema==4.1.2
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pylint==2.3.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.2
SQLAlchemy==1.3
systemd-python==234
terminado==0.12.1
urllib3==1.26.2
Werkzeug==0.15.5
WTForms==2.3.3
python-dotenv==0.19.1
python-jose-cryptodome==1.3.2